vidc: fixes squashed from caf
* vidc: Send port reconfiguration for interlaced clips.

Component is not getting port reconfiguration request
from driver for an interlaced clip. Fix it by
modifying the condition for triggering the
reconfiguration request.

CRs-Fixed: 352456
Change-Id: I65ce5987f31915f45a13767965918b61e32db7a7
Signed-off-by: Rajeshwar Kurapaty <rkurapat@codeaurora.org>
(cherry picked from commit c264eed)
Signed-off-by: Darescu Ionut <johnny@esi.ro>

* vidc: Free shared memory when video core is in error.

When the video core is in error, the core times out.
At this instant, all the client buffers are freed,
but the shared memory is not freed. Free the shared
memory in this case.

CRs-Fixed: 357990
Signed-off-by: Deepak Kotur <dkotur@codeaurora.org>
(cherry picked from commit 17ca368)

Change-Id: I7695d4b101d8b769276a73100de9cb2495d1ccb9
Signed-off-by: Gopal G Goberu <ggober@codeaurora.org>
Signed-off-by: Darescu Ionut <johnny@esi.ro>